removing CSV codec and integrating with parquet module
