+++ gradle/missing-javadoc.gradle
@@ -165,7 +165,6 @@ configure([
   project(":plugins:engine-datafusion"), //TODO
   project(":server"),
   project(":modules:parquet-data-format"),
-  project(":plugins:dataformat-csv"), //TODO
 ]) {
   project.tasks.withType(MissingJavadocTask) {
     isExcluded = true
