+++ modules/parquet-data-format/src/main/resources/META-INF/services/org.opensearch.vectorized.execution.search.spi.DataSourceCodec
@@ -1,1 +1,1 @@
-org.opensearch.datafusion.csv.CsvDataSourceCodec
+com.parquet.parquetdataformat.engine.read.ParquetDataSourceCodec
