--- modules/parquet-data-format/src/main/rust/src/csv_exec.rs
@@ -1,24 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- */
-
-use anyhow::Result;
-
-/// Csv-specific execution utilities - placeholder implementation
-pub struct CsvExecutor;
-
-impl CsvExecutor {
-    pub fn new() -> Self {
-        Self
-    }
-    
-    /// Create a listing table for Csv files - placeholder
-    pub async fn create_csv_table(
-        &self,
-        table_path: &str,
-    ) -> Result<u64> {
-        // Placeholder implementation
-        log::info!("Creating csv table for path: {}", table_path);
-        Ok(1) // Return dummy table ID
-    }
-}
