+++ modules/parquet-data-format/src/main/rust/src/parquet_exec.rs
@@ -1,0 +1,24 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+use anyhow::Result;
+
+/// Parquet-specific execution utilities - placeholder implementation
+pub struct ParquetExecutor;
+
+impl ParquetExecutor {
+    pub fn new() -> Self {
+        Self
+    }
+
+    /// Create a listing table for Parquet files - placeholder
+    pub async fn create_parquet_table(
+        &self,
+        table_path: &str,
+    ) -> Result<u64> {
+        // Placeholder implementation
+        log::info!("Creating parquet table for path: {}", table_path);
+        Ok(1) // Return dummy table ID
+    }
+}
