--- plugins/dataformat-csv/jni/src/context.rs
@@ -1,70 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- */
-
-use datafusion::prelude::*;
-use datafusion::execution::context::SessionContext;
-use std::collections::HashMap;
-use std::sync::Arc;
-use anyhow::Result;
-
-/// Manages DataFusion session contexts
-pub struct SessionContextManager {
-    contexts: HashMap<*mut SessionContext, Arc<SessionContext>>,
-    next_runtime_id: u64,
-}
-
-impl SessionContextManager {
-    pub fn new() -> Self {
-        Self {
-            contexts: HashMap::new(),
-            next_runtime_id: 1,
-        }
-    }
-
-    pub async fn register_directory(
-        &mut self,
-        table_name: &str,
-        directory_path: &str,
-        options: HashMap<String, String>,
-    ) -> Result<u64> {
-        // Placeholder implementation - would register csv directory as table
-        log::info!("Registering directory: {} at path: {} with options: {:?}",
-                   table_name, directory_path, options);
-
-        let runtime_id = self.next_runtime_id;
-        self.next_runtime_id += 1;
-        Ok(runtime_id)
-    }
-
-    pub async fn create_session_context(
-        &mut self,
-        config: HashMap<String, String>,
-    ) -> Result<*mut SessionContext> {
-        // Create actual DataFusion session context
-        let mut session_config = SessionConfig::new();
-
-        // Apply configuration options
-        if let Some(batch_size) = config.get("batch_size") {
-            if let Ok(size) = batch_size.parse::<usize>() {
-                session_config = session_config.with_batch_size(size);
-            }
-        }
-
-        let ctx = Arc::new(SessionContext::new_with_config(session_config));
-        let ctx_ptr = Arc::as_ptr(&ctx) as *mut SessionContext;
-
-        self.contexts.insert(ctx_ptr, ctx);
-
-        Ok(ctx_ptr)
-    }
-
-    pub async fn close_session_context(&mut self, ctx_ptr: *mut SessionContext) -> Result<()> {
-        self.contexts.remove(&ctx_ptr);
-        Ok(())
-    }
-
-    pub fn get_context(&self, ctx_ptr: *mut SessionContext) -> Option<&Arc<SessionContext>> {
-        self.contexts.get(&ctx_ptr)
-    }
-}
