--- plugins/dataformat-csv/jni/src/stream.rs
@@ -1,43 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- */
-
-use anyhow::Result;
-use serde_json;
-
-/// Wrapper for DataFusion record batch streams - placeholder implementation
-pub struct RecordBatchStreamWrapper {
-    batch_count: u32,
-    is_placeholder: bool,
-}
-
-impl RecordBatchStreamWrapper {
-    pub fn new_placeholder() -> Self {
-        Self { 
-            batch_count: 0,
-            is_placeholder: true,
-        }
-    }
-    
-    pub async fn next_batch(&mut self) -> Result<Option<String>> {
-        // Return placeholder data for first few calls, then None
-        if self.is_placeholder {
-            if self.batch_count < 2 {
-                self.batch_count += 1;
-                let placeholder_data = serde_json::json!({
-                    "rows": [
-                        {"id": self.batch_count, "name": format!("placeholder_row_{}", self.batch_count)}
-                    ],
-                    "num_rows": 1,
-                    "num_columns": 2
-                });
-                Ok(Some(serde_json::to_string(&placeholder_data)?))
-            } else {
-                Ok(None) // End of stream
-            }
-        } else {
-            // Real implementation would go here
-            Ok(None)
-        }
-    }
-}
