--- plugins/dataformat-csv/jni/src/substrait.rs
@@ -1,37 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- */
-
-use datafusion::execution::context::SessionContext;
-use crate::stream::RecordBatchStreamWrapper;
-use anyhow::Result;
-
-/// Executes Substrait query plans
-pub struct SubstraitExecutor;
-
-impl SubstraitExecutor {
-    pub fn new() -> Self {
-        Self
-    }
-    
-    pub async fn execute_plan(
-        &self,
-        session_context_ptr: *mut SessionContext,
-        substrait_plan_bytes: &[u8],
-    ) -> Result<*mut RecordBatchStreamWrapper> {
-        // Placeholder implementation - would normally:
-        // 1. Parse Substrait plan from substrait_plan_bytes
-        // 2. Convert to DataFusion logical plan using datafusion-substrait
-        // 3. Execute using the session context
-        // 4. Return actual record batch stream
-        
-        log::info!("Executing Substrait plan with {} bytes for session: {:?}", 
-                   substrait_plan_bytes.len(), session_context_ptr);
-        
-        // For now, return a placeholder stream
-        let wrapper = RecordBatchStreamWrapper::new_placeholder();
-        let wrapper_ptr = Box::into_raw(Box::new(wrapper));
-        
-        Ok(wrapper_ptr)
-    }
-}
