--- plugins/dataformat-csv/src/main/java/org/opensearch/datafusion/csv/CsvRecordBatchStream.java
@@ -1,119 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-package org.opensearch.datafusion.csv;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.opensearch.vectorized.execution.search.spi.RecordBatchStream;
-
-import java.util.concurrent.CompletableFuture;
-
-/**
- * TODO : this need not be here - nothing specific to CSV - move to LIB ?
- * Native implementation of RecordBatchStream that wraps a JNI stream pointer.
- * This class provides a Java interface over native DataFusion record batches.
- */
-public class CsvRecordBatchStream implements RecordBatchStream {
-
-    private static final Logger logger = LogManager.getLogger(CsvRecordBatchStream.class);
-
-    private final long nativeStreamPtr;
-    private volatile boolean closed = false;
-    private volatile boolean hasNextCached = false;
-    private volatile boolean hasNextValue = false;
-
-    /**
-     * Creates a new CsvRecordBatchStream wrapping the given native stream pointer.
-     *
-     * @param nativeStreamPtr Pointer to the native DataFusion RecordBatch stream
-     */
-    public CsvRecordBatchStream(long nativeStreamPtr) {
-        if (nativeStreamPtr == 0) {
-            throw new IllegalArgumentException("Invalid native stream pointer");
-        }
-        this.nativeStreamPtr = nativeStreamPtr;
-        logger.debug("Created CsvRecordBatchStream with pointer: {}", nativeStreamPtr);
-    }
-
-    @Override
-    public Object getSchema() {
-        return "CsvSchema"; // Placeholder
-    }
-
-    @Override
-    public CompletableFuture<Object> next() {
-        // PlaceholderImpl
-        return CompletableFuture.supplyAsync(() -> {
-            if (closed) {
-                return null;
-            }
-
-            try {
-                // Get the next batch from native code
-                String batch = nativeNextBatch(nativeStreamPtr);
-
-                // Reset cached hasNext value since we consumed a batch
-                hasNextCached = false;
-
-                logger.trace("Retrieved next batch from stream pointer: {}", nativeStreamPtr);
-                return batch;
-            } catch (Exception e) {
-                logger.error("Error getting next batch from stream", e);
-                return null;
-            }
-        });
-    }
-
-    @Override
-    public boolean hasNext() {
-        // Placeholder impl
-        if (closed) {
-            return false;
-        }
-
-        if (hasNextCached) {
-            return hasNextValue;
-        }
-
-        try {
-            // Check if there's a next batch available
-            // This is a simplified implementation - in practice, you might want to
-            // peek at the stream without consuming the batch
-            String nextBatch = nativeNextBatch(nativeStreamPtr);
-            hasNextValue = (nextBatch != null);
-            hasNextCached = true;
-
-            logger.trace("hasNext() = {} for stream pointer: {}", hasNextValue, nativeStreamPtr);
-            return hasNextValue;
-        } catch (Exception e) {
-            logger.error("Error checking for next batch in stream", e);
-            return false;
-        }
-    }
-
-    @Override
-    public void close() {
-        if (!closed) {
-            logger.debug("Closing CsvRecordBatchStream with pointer: {}", nativeStreamPtr);
-            try {
-                nativeCloseStream(nativeStreamPtr);
-                closed = true;
-                logger.debug("Successfully closed CsvRecordBatchStream");
-            } catch (Exception e) {
-                logger.error("Error closing CsvRecordBatchStream", e);
-                throw e;
-            }
-        }
-    }
-
-    // Native method declarations
-    private static native String nativeNextBatch(long streamPtr);
-
-    private static native void nativeCloseStream(long streamPtr);
-}
