--- plugins/dataformat-csv/src/main/java/org/opensearch/datafusion/csv/package-info.java
@@ -1,13 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-/**
- * CSV data format implementation for DataFusion integration.
- * Provides CSV file reading capabilities through DataFusion query engine.
- */
-package org.opensearch.datafusion.csv;
