--- plugins/dataformat-csv/src/main/resources/META-INF/services/org.opensearch.vectorized.execution.search.spi.DataSourceCodec
@@ -1,1 +1,0 @@
-org.opensearch.datafusion.csv.CsvDataSourceCodec
