--- plugins/dataformat-csv/src/main/resources/plugin-descriptor.properties
@@ -1,7 +1,0 @@
-# Plugin descriptor for CSV data format plugin
-description=CSV data format plugin for OpenSearch DataFusion
-version=${project.version}
-name=dataformat-csv
-classname=org.opensearch.datafusion.csv.CsvDataFormatPlugin
-java.version=${versions.java}
-opensearch.version=${opensearch_version}
